pyproject.toml
[tool.poetry]
name = "adearn"
version = "0.1.0"
description = ""
authors = ["braviss <doctype.developer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
factory-boy = "^3.3.0"
pillow = "^10.2.0"
django-tinymce = "^3.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
